Improve augmentation in ckeditor5-special-characters.
