+++ packages/ckeditor5-special-characters/src/index.ts
@@ -14,3 +14,7 @@ export { default as SpecialCharactersMathematical } from './specialcharactersmat
 export { default as SpecialCharactersLatin } from './specialcharacterslatin';
 export { default as SpecialCharactersEssentials } from './specialcharactersessentials';
 export { default as SpecialCharactersCurrency } from './specialcharacterscurrency';
+
+export { SpecialCharactersConfig } from './specialcharactersconfig';
+
+import './augmentation';
